fix and name

+++ 3822B1FI1/1_gelu_omp/beresnev_a/gelu_omp.cpp
@@ -8,7 +8,7 @@ AlignedVector GeluOMP(const AlignedVector& input) {
     #pragma omp parallel for
     for (std::size_t i = 0; i < input.size(); ++i) {
         float x = input[i];
-        output[i] = 0.5f * x * (1.0f + std::tanhf(std::sqrt(2.0f / M_PI) * (x + 0.044715f * x * x * x)));
+        output[i] = 0.5f * x * (1.0f + std::tanh(std::sqrt(2.0f / M_PI) * (x + 0.044715f * x * x * x)));
     }
 
     return output;
